requirements.txt
# 银行管理系统依赖包
# 数据库连接
mysql-connector-python==8.0.33

# 表格美化显示
tabulate==0.9.0
